API

QuadraticOptimizer.optimize_qfm! — Method

optimize_qfm!(f, ps::Vector{<:SVector{D, <:Real}}, fs::Vector{<:Real}, n::Integer)

Optimize a function f using the Quadratic Fitting Method (QFM).

Arguments

f: The objective function to be optimized.

ps: A vector of points in $\mathbb{R}^d$ where f has been evaluated. This vector will be updated in-place during the optimization process.

fs: A vector of function values corresponding to the points in ps. This vector will be updated in-place during the optimization process.

n: The number of optimizing iterations. After execution, the length of ps will be m + n, where m = length(ps) before execution.

Note

In each step of the QFM, the last m points from ps and fs are used to fit a quadratic function. The method iteratively refines the points and function values, extending ps and fs with n additional points resulting from the optimization process.

Examples

julia> using QuadraticOptimizer, StaticArrays, Random

julia> Random.seed!(42);

julia> f(p) = p[1]^2 + sin(p[1]) + 1.5p[2]^2 + sinh(p[2]) - p[1]*p[2]/5
f (generic function with 1 method)

julia> ps_init = [@SVector rand(2) for _ in 1:10]
10-element Vector{SVector{2, Float64}}:
 [0.6293451231426089, 0.4503389405961936]
 [0.47740714343281776, 0.7031298490032014]
 [0.6733461456394962, 0.16589443479313404]
 [0.6134782250008441, 0.6683403279577278]
 [0.4570310908017041, 0.2993652953937611]
 [0.6611433726193705, 0.6394313620423493]
 [0.34264792290134793, 0.2678704383989201]
 [0.515871408349502, 0.09002301604339691]
 [0.27265744579429385, 0.191562202596938]
 [0.4235912564725989, 0.4847023673932017]

julia> ps = copy(ps_init);

julia> fs = f.(ps);

julia> optimize_qfm!(f, ps, fs, 20);

QuadraticOptimizer.optimize_qfm — Method

optimize_qfm(f, ps::Vector{<:SVector{D, <:Real}}, n::Integer)

Optimize a function f using the Quadratic Fitting Method (QFM).

Arguments

f: The objective function to be optimized.

ps: A vector of points in $\mathbb{R}^d$ where f has been evaluated.

n: The number of optimizing iterations. After execution, the length of ps will be m + n, where m = length(ps) before execution.

Note

In each step of the QFM, the last m points from ps and fs are used to fit a quadratic function. The method iteratively refines the points and function values, extending ps and fs with n additional points resulting from the optimization process.

Examples

julia> using QuadraticOptimizer, StaticArrays, Random

julia> Random.seed!(42);

julia> f(p) = p[1]^2 + sin(p[1]) + 1.5p[2]^2 + sinh(p[2]) - p[1]*p[2]/5
f (generic function with 1 method)

julia> ps_init = [@SVector rand(2) for _ in 1:10]
10-element Vector{SVector{2, Float64}}:
 [0.6293451231426089, 0.4503389405961936]
 [0.47740714343281776, 0.7031298490032014]
 [0.6733461456394962, 0.16589443479313404]
 [0.6134782250008441, 0.6683403279577278]
 [0.4570310908017041, 0.2993652953937611]
 [0.6611433726193705, 0.6394313620423493]
 [0.34264792290134793, 0.2678704383989201]
 [0.515871408349502, 0.09002301604339691]
 [0.27265744579429385, 0.191562202596938]
 [0.4235912564725989, 0.4847023673932017]

julia> ps, fs = optimize_qfm(f, ps_init, 20);

QuadraticOptimizer.optimize_qfm — Method

optimize_qfm(f, ps::Vector{<:SVector{D, <:Real}}, fs::Vector{<:Real}, n::Integer)

Optimize a function f using the Quadratic Fitting Method (QFM).

Arguments

f: The objective function to be optimized.

ps: A vector of points in $\mathbb{R}^d$ where f has been evaluated.

fs: A vector of function values corresponding to the points in ps.

n: The number of optimizing iterations. After execution, the length of ps will be m + n, where m = length(ps) before execution.

Note

In each step of the QFM, the last m points from ps and fs are used to fit a quadratic function. The method iteratively refines the points and function values, extending ps and fs with n additional points resulting from the optimization process.

Examples

julia> using QuadraticOptimizer, StaticArrays, Random

julia> Random.seed!(42);

julia> f(p) = p[1]^2 + sin(p[1]) + 1.5p[2]^2 + sinh(p[2]) - p[1]*p[2]/5
f (generic function with 1 method)

julia> ps_init = [@SVector rand(2) for _ in 1:10]
10-element Vector{SVector{2, Float64}}:
 [0.6293451231426089, 0.4503389405961936]
 [0.47740714343281776, 0.7031298490032014]
 [0.6733461456394962, 0.16589443479313404]
 [0.6134782250008441, 0.6683403279577278]
 [0.4570310908017041, 0.2993652953937611]
 [0.6611433726193705, 0.6394313620423493]
 [0.34264792290134793, 0.2678704383989201]
 [0.515871408349502, 0.09002301604339691]
 [0.27265744579429385, 0.191562202596938]
 [0.4235912564725989, 0.4847023673932017]

julia> ps, fs = optimize_qfm(f, ps_init, f.(ps_init), 20);

QuadraticOptimizer.optimize_qim! — Method

optimize_qim!(f, xs::Vector{<:Real}, fs::Vector{<:Real}, n::Integer)

Optimize a function f using the Quadratic Interpolation Method (QIM).

Arguments

f: The objective function to be optimized.

xs: A vector of points in $\mathbb{R}^d$ where f has been evaluated. This vector will be updated in-place during the optimization process.

fs: A vector of function values corresponding to the points in xs. This vector will be updated in-place during the optimization process.

n: The number of optimizing iterations. After execution, the length of xs will be m + n, where m = length(xs) before execution.

Note

In each step of the QIM, the last 3 points from xs and fs are used to interpolate with a quadratic function. The method iteratively refines the points and function values, extending xs and fs with n additional points resulting from the optimization process.

Examples

julia> using QuadraticOptimizer

julia> f(x) = sin(x) + x^2/10  # Function to minimize
f (generic function with 1 method)

julia> xs_init = [1.2, 0.1, -2.2]  # Initial points (3 points are required to construct parabola)
3-element Vector{Float64}:
  1.2
  0.1
 -2.2

julia> xs = copy(xs_init);

julia> fs = f.(xs);

julia> optimize_qim!(f, xs, fs, 20);

QuadraticOptimizer.optimize_qim! — Method

optimize_qim!(f, ps::Vector{<:SVector{D, <:Real}}, fs::Vector{<:Real}, n::Integer)

Optimize a function f using the Quadratic Interpolation Method (QIM).

Arguments

f: The objective function to be optimized.

ps: A vector of points in $\mathbb{R}^d$ where f has been evaluated. This vector will be updated in-place during the optimization process.

fs: A vector of function values corresponding to the points in ps. This vector will be updated in-place during the optimization process.

n: The number of optimizing iterations. After execution, the length of ps will be m + n, where m = length(ps) before execution.

Note

In each step of the QIM, the last m (==(D*(D+1)/2)) points from ps and fs are used to interpolate with a quadratic function. The method iteratively refines the points and function values, extending ps and fs with n additional points resulting from the optimization process.

Examples

julia> using QuadraticOptimizer, StaticArrays, Random

julia> Random.seed!(42);

julia> f(p) = p[1]^2 + sin(p[1]) + 1.5p[2]^2 + sinh(p[2]) - p[1]*p[2]/5
f (generic function with 1 method)

julia> ps_init = [@SVector rand(2) for _ in 1:6]
6-element Vector{SVector{2, Float64}}:
 [0.6293451231426089, 0.4503389405961936]
 [0.47740714343281776, 0.7031298490032014]
 [0.6733461456394962, 0.16589443479313404]
 [0.6134782250008441, 0.6683403279577278]
 [0.4570310908017041, 0.2993652953937611]
 [0.6611433726193705, 0.6394313620423493]

julia> ps = copy(ps_init);

julia> fs = f.(ps);

julia> optimize_qim!(f, ps, fs, 20);

QuadraticOptimizer.optimize_qim — Method

optimize_qfm(f, xs::Vector{<:Real}, n::Integer)

Optimize a function f using the Quadratic Interpolation Method (QIM).

Arguments

f: The objective function to be optimized.

xs: A vector of points in $\mathbb{R}^d$ where f has been evaluated.

n: The number of optimizing iterations. After execution, the length of xs will be m + n, where m = length(xs) before execution.

Note

In each step of the QIM, the last 3 points from xs and fs are used to interpolate with a quadratic function. The method iteratively refines the points and function values, extending xs and fs with n additional points resulting from the optimization process.

Examples

julia> using QuadraticOptimizer

julia> f(x) = sin(x) + x^2/10  # Function to minimize
f (generic function with 1 method)

julia> xs_init = [1.2, 0.1, -2.2]  # Initial points (3 points are required to construct parabola)
3-element Vector{Float64}:
  1.2
  0.1
 -2.2

julia> xs = copy(xs_init)  # Keep initial points
3-element Vector{Float64}:
  1.2
  0.1
 -2.2

julia> xs, fs = optimize_qim(f, xs, 10);  # Optimize 10 steps

QuadraticOptimizer.optimize_qim — Method

optimize_qim(f, xs::Vector{<:Real}, fs::Vector{<:Real}, n::Integer)

Optimize a function f using the Quadratic Interpolation Method (QIM).

Arguments

f: The objective function to be optimized.

xs: A vector of points in $\mathbb{R}^d$ where f has been evaluated.

fs: A vector of function values corresponding to the points in xs.

n: The number of optimizing iterations. After execution, the length of xs will be m + n, where m = length(xs) before execution.

Note

In each step of the QIM, the last 3 points from xs and fs are used to interpolate with a quadratic function. The method iteratively refines the points and function values, extending xs and fs with n additional points resulting from the optimization process.

Examples

julia> using QuadraticOptimizer

julia> f(x) = sin(x) + x^2/10  # Function to minimize
f (generic function with 1 method)

julia> xs_init = [1.2, 0.1, -2.2]  # Initial points (3 points are required to construct parabola)
3-element Vector{Float64}:
  1.2
  0.1
 -2.2

julia> xs, fs = optimize_qim(f, xs_init, f.(xs_init), 10);  # Optimize 10 steps

QuadraticOptimizer.optimize_qim — Method

optimize_qfm(f, ps::Vector{<:SVector{D, <:Real}}, n::Integer)

Optimize a function f using the Quadratic Interpolation Method (QIM).

Arguments

f: The objective function to be optimized.

ps: A vector of points in $\mathbb{R}^d$ where f has been evaluated.

n: The number of optimizing iterations. After execution, the length of ps will be m + n, where m = length(ps) before execution.

Note

In each step of the QIM, the last m (==(D*(D+1)/2)) points from ps and fs are used to interpolate with a quadratic function. The method iteratively refines the points and function values, extending ps and fs with n additional points resulting from the optimization process.

Examples

julia> using QuadraticOptimizer, StaticArrays, Random

julia> Random.seed!(42);

julia> f(p) = p[1]^2 + sin(p[1]) + 1.5p[2]^2 + sinh(p[2]) - p[1]*p[2]/5
f (generic function with 1 method)

julia> ps_init = [@SVector rand(2) for _ in 1:6]
6-element Vector{SVector{2, Float64}}:
 [0.6293451231426089, 0.4503389405961936]
 [0.47740714343281776, 0.7031298490032014]
 [0.6733461456394962, 0.16589443479313404]
 [0.6134782250008441, 0.6683403279577278]
 [0.4570310908017041, 0.2993652953937611]
 [0.6611433726193705, 0.6394313620423493]

julia> ps, fs = optimize_qfm(f, ps_init, 20);

QuadraticOptimizer.optimize_qim — Method

optimize_qim(f, ps::Vector{<:SVector{D, <:Real}}, fs::Vector{<:Real}, n::Integer)

Optimize a function f using the Quadratic Interpolation Method (QIM).

Arguments

f: The objective function to be optimized.

ps: A vector of points in $\mathbb{R}^d$ where f has been evaluated.

fs: A vector of function values corresponding to the points in ps.

n: The number of optimizing iterations. After execution, the length of ps will be m + n, where m = length(ps) before execution.

Note

In each step of the QIM, the last m (==(D*(D+1)/2)) points from ps and fs are used to interpolate with a quadratic function. The method iteratively refines the points and function values, extending ps and fs with n additional points resulting from the optimization process.

Examples

julia> using QuadraticOptimizer, StaticArrays, Random

julia> Random.seed!(42);

julia> f(p) = p[1]^2 + sin(p[1]) + 1.5p[2]^2 + sinh(p[2]) - p[1]*p[2]/5
f (generic function with 1 method)

julia> ps_init = [@SVector rand(2) for _ in 1:6]
6-element Vector{SVector{2, Float64}}:
 [0.6293451231426089, 0.4503389405961936]
 [0.47740714343281776, 0.7031298490032014]
 [0.6733461456394962, 0.16589443479313404]
 [0.6134782250008441, 0.6683403279577278]
 [0.4570310908017041, 0.2993652953937611]
 [0.6611433726193705, 0.6394313620423493]

julia> ps, fs = optimize_qim(f, ps_init, f.(ps_init), 20);

QuadraticOptimizer.quadratic_fitting — Method

quadratic_fitting(ps::AbstractVector{<:StaticVector{D, <:Real}}, fs::AbstractVector{<:Real}) where D

Calculate quadratic fitting based on input values.

Examples

julia> using QuadraticOptimizer, StaticArrays, Random

julia> Random.seed!(42);

julia> q = Quadratic{2}([2,1,3], [1,2], 2)
Quadratic{2, 3, Int64}([2, 1, 3], [1, 2], 2)

julia> ps = [@SVector rand(2) for _ in 1:6]
6-element Vector{SVector{2, Float64}}:
 [0.6293451231426089, 0.4503389405961936]
 [0.47740714343281776, 0.7031298490032014]
 [0.6733461456394962, 0.16589443479313404]
 [0.6134782250008441, 0.6683403279577278]
 [0.4570310908017041, 0.2993652953937611]
 [0.6611433726193705, 0.6394313620423493]

julia> quadratic_fitting(vcat(ps, ps), vcat(q.(ps).-1, q.(ps).+1))
Quadratic{2, 3, Float64}([2.0000000002328306, 0.9999999999890861, 2.999999999989086], [0.9999999997671694, 2.000000000021828], 2.0000000000436557)

QuadraticOptimizer.quadratic_interpolation — Method

quadratic_interpolation(ps::AbstractVector{<:StaticVector{D, <:Real}}, fs::AbstractVector{<:Real}) where D

Calculate quadratic interpolation based on input values.

Examples

julia> using QuadraticOptimizer, StaticArrays, Random

julia> Random.seed!(42);

julia> q = Quadratic{2}([2,1,3], [1,2], 2)
Quadratic{2, 3, Int64}([2, 1, 3], [1, 2], 2)

julia> ps = [@SVector rand(2) for _ in 1:6]
6-element Vector{SVector{2, Float64}}:
 [0.6293451231426089, 0.4503389405961936]
 [0.47740714343281776, 0.7031298490032014]
 [0.6733461456394962, 0.16589443479313404]
 [0.6134782250008441, 0.6683403279577278]
 [0.4570310908017041, 0.2993652953937611]
 [0.6611433726193705, 0.6394313620423493]

julia> quadratic_interpolation(ps, q.(ps))
Quadratic{2, 3, Float64}([1.9999999999999962, 0.9999999999999407, 3.000000000000005], [1.0000000000000533, 2.000000000000017], 1.9999999999999492)